lab4/requirements.txt
AWSIoTPythonSDK==1.5.2
black==23.1.0
boto3==1.26.100
botocore==1.29.100
click==8.1.3
jmespath==1.0.1
joblib==1.2.0
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
platformdirs==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
